Never reject with undefined
That's like `throw`ing undefined, there is no context and it's impossible for users to be aware of what happened.

Even doing something simple like rejecting with an `Error` helps here. 

I'm not sure what's the point of even rejecting here semantically (vs. not fulfilling) - it doesn't make a lot of sense and introduces abort and control at a distance semantics which are highly discouraged when using promises.

Cheers and thanks for the library.